Refactor app_secrets.hpp
This file is no longer directly provided.

Instead an example file (app_secrets.example.hpp) is provided which the
user can copy into app_secrets.hpp which is now being ignored.

This should help with avoiding accidentally commiting changes to the
secrets file, as well as make sure API configuration is not lost on
checkout.